Adjustments to movie action buttons || Adjustments to Agenda info
